Move queue-related logic to new queue lib
