+++ libs/queue/src/error.rs
