+++ libs/queue/src/lib.rs
@@ -1,7 +1,9 @@
-use crate::{
-    error::Error,
-    job::{PostgresJob, PostgresJobStatus},
-};
+mod error;
+pub mod job;
+pub mod runner;
+
+use crate::error::Error;
+use crate::job::{PostgresJob, PostgresJobStatus};
 use serde::{Deserialize, Serialize};
 use sqlx::{self, types::Json, PgPool};
 use std::fmt::Debug;
